refactor: move Activity to Activity package
