+++ src/main/java/com/alkemy/ong/domain/activity/ActivityGateway.java
@@ -1,7 +1,5 @@
 package com.alkemy.ong.domain.activity;
 
-import com.alkemy.ong.domain.model.Activity;
-
 public interface ActivityGateway {
   Activity save(Activity activity);
 }
